created bizcard file and moved all related files
