+++ client/src/tanner.css
@@ -11,28 +11,3 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
-
-#location-container{
-  display: flex; 
-  flex-flow: row wrap;
-  justify-content: center;
-}
-
-#location-cell{
-  background-color: lightgreen;
-  width: 300px;
-  margin: 10px;
-  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.11), 0 4px 15px 0 rgba(0, 0, 0, 0.103)
-}
-
-#business-image{
-  margin-top: 15px;
-  width: 250px;
-  min-height: 100%;
-  object-fit: cover;
-}
-
-#location-png{
-  height:20px;
-  width:15px;
-}
